api/types/swarm: move `NodeListOptions` to client mod
